backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
httpx==0.27.0

# Redis & jobs
redis==5.0.4
celery==5.4.0
celery-redbeat==2.3.3
flower==2.0.1

# Storage
boto3==1.34.120
Pillow==10.3.0

# Push notifications
pywebpush==2.0.0

# Validation & serialization
pydantic==2.7.1
pydantic-settings==2.3.1
email-validator==2.1.1

# Utilities
python-dotenv==1.0.1
tenacity==8.3.0


